Minor core changes
BotExceptions can now be raised during an active message. Getting a file from the cache sets the access time, and extensions can be added to some utility functions. 

Exceptions now also have the ability to edit a given message to the given error message. However, this change may or may not stay depending on how well it can be worked around.