Add voice channel joining and leaving
You can now manually pull the bot into and out of voice channels. Makes sound tag startup a little faster, as the time used up to connect to the channel can be used to read the tag itself.
Also, added a wrapper for getting channels in a similar fashion to getting members.