Feedback, toggling commands, and hotloading plugins
It's all packaged together! Server owners can send feedback to the bot owners (which can be blacklisted or disabled from the bot owner's side).
For speed of development, a subcommand was created just to reload external plugins (and load new ones if found). Not recommended for use in a non-development bot.