Add invite subcommand, and subcommand toggling
The bot can now give you an invite to itself via the `!base invite` command. There is also a detailed mode that shows the user exactly what each permission is used for.

Subcommands can now be toggled. As an example, tag editing can be disabled with `!mod toggle tag edit` or `!mod toggle tag 6` if 6 is the 6th subcommand of `tag`, given by `!help tag`.