In-chat update and per-server timezone configuration
Servers now have the option to set their own timezone via UTC offset. Used for the updated GDQ plugin.

There is now an in-chat update menu. This removes the need to use the external update.py script whenever an update is needed. Furthermore, this should update discord.py when the core is updated.