Single command mode
Specialized bots that rely on one plugin used to require the base for every command invocation. No more! Now, a fallback base can be set in the configuration file.

For example, if the single command is "tag", issuing the command "!create foo bar" gets internally converted to "!tag create foo bar" because "create" is not a valid base command.