Maintenance mode, documentation cleanup, ...
Maintenance mode allows the bot to not respond to any messages, or show a global error.

Lots of documentation was added for message types and argument types.

Response handling has been restructured so that plugins can take advantage of things like the interactive menu.

The permissions model is now an IntEnum instead of just an int.

There is now a utility function in the data module for getting a user's elevation level.

Exceptions can now be configured to not display the "editable" message.